Fix warnings and other small issues.

+++ formula-ml/core/util/Util.cpp
@@ -4,5 +4,5 @@
 
 long util::current_time() {
     auto now = std::chrono::system_clock::now().time_since_epoch();
-    return std::chrono::duration_cast<std::chrono::milliseconds>(now).count();
+    return (long) std::chrono::duration_cast<std::chrono::milliseconds>(now).count();
 }
